F01: Add implementation of PrivilegeService
